% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseiii.R
\name{read_phaseiii}
\alias{read_phaseiii}
\title{Read a Phase III deployment workbook}
\usage{
read_phaseiii(deployid)
}
\arguments{
\item{deployid}{Deployment id e.g. "bw180905-53"}
}
\value{
A \code{tibble} with columns \code{deployid}, \code{motionlessid},
\code{motionlessstart}, \code{motionlessend}, and \code{duration_s}
}
\description{
Read a Phase III deployment workbook
}
