% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tsviewer.R
\name{run_tsviewer}
\alias{run_tsviewer}
\title{Run the time series viewer}
\usage{
run_tsviewer(prh, resume = FALSE)
}
\arguments{
\item{prh}{A PRH object. See catsr::read_nc().}

\item{resume}{Resume progress flag. If \code{TRUE}, searches for deployment
workbook in the Phase III shared Google Drive folder.}
}
\description{
Run the time series viewer
}
