% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tsviewer.R
\name{run_tsviewer}
\alias{run_tsviewer}
\title{Run the time series viewer}
\usage{
run_tsviewer(prh)
}
\arguments{
\item{prh}{A PRH object. See catsr::read_nc().}
}
\description{
Run the time series viewer
}
